// generated/clients/javascript/services/uigf-game-genshin-sdk-secure/src/api/secureApi/options.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

import { OperationOptions } from "@azure-rest/core-client";

/** Optional parameters. */
export interface SecureApiGetRedDotsOptionalParams extends OperationOptions {
  rpcAppId?: string;
  rpcAppVersion?: string;
  rpcChannelId?: string;
  rpcChannelVersion?: string;
  rpcClientType?: string;
  rpcComboVersion?: string;
  rpcDeviceFingerprint?: string;
  rpcDeviceId?: string;
  rpcDeviceModel?: string;
  rpcDeviceName?: string;
  rpcGameBiz?: string;
  rpcGoodsThirdParty?: string;
  rpcLanguage?: string;
  rpcMdkVersion?: string;
  rpcPaymentVersion?: string;
  rpcSdkVersion?: string;
  rpcSubChannelId?: string;
  rpcSystemVersion?: string;
}

/** Optional parameters. */
export interface SecureApiCompareProtocolVersionOptionalParams extends OperationOptions {
  rpcAppId?: string;
  rpcAppVersion?: string;
  rpcChannelId?: string;
  rpcChannelVersion?: string;
  rpcClientType?: string;
  rpcComboVersion?: string;
  rpcDeviceFingerprint?: string;
  rpcDeviceId?: string;
  rpcDeviceModel?: string;
  rpcDeviceName?: string;
  rpcGameBiz?: string;
  rpcGoodsThirdParty?: string;
  rpcLanguage?: string;
  rpcMdkVersion?: string;
  rpcPaymentVersion?: string;
  rpcSdkVersion?: string;
  rpcSubChannelId?: string;
  rpcSystemVersion?: string;
}